% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combine_nimble.R
\name{combine_nimble}
\alias{combine_nimble}
\title{Combine Nimble output}
\usage{
combine_nimble(Y, sims_in, like_call_in, priors_in, sigma_in = 1)
}
\arguments{
\item{Y}{data}

\item{sims_in}{list of mcmc sims generated from run_nimble()}

\item{like_call_in}{list of likelihood calls associated with each model}

\item{priors_in}{vector of prior probabilities}

\item{sigma_in}{known sigma value}
}
\value{
list of posterior probabilities
}
\description{
Combine Nimble output
}
\examples{
\dontrun{
combine_nimble()
}
}
