% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{coords}
\alias{coords}
\title{Coordinates}
\format{
\subsection{\code{coords}}{

A data frame with 1000 rows and 2 columns:
\describe{
\item{x}{longitude}
\item{y}{ latitude}
}
}
}
\source{
Obtained from National Bridge Inventory database
}
\usage{
coords
}
\description{
Coordinates for potential bat roosts
}
\keyword{datasets}
