% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_landuse.R
\name{extract_landuse}
\alias{extract_landuse}
\title{Extract land use from USGS NLCD}
\usage{
extract_landuse(lat, long, SiteName, buffer = 100)
}
\arguments{
\item{lat}{Latitude for point}

\item{long}{Longitude for point}

\item{SiteName}{Name associated with point}

\item{buffer}{Distance around point in meters.}
}
\value{
tibble with land use type and proportion
}
\description{
Extract land use from USGS NLCD
}
\examples{
\dontrun{
rgee::ee_Initialize()
extract_landuse(45.676998, -111.042931, 'Bozeman', buffer = 100)
}
}
