% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prior_auto.R
\name{prior_auto}
\alias{prior_auto}
\title{1D Prior Elicitation with automatic features}
\usage{
prior_auto()
}
\value{
opens shiny app
}
\description{
1D Prior Elicitation with automatic features
}
\examples{
\dontrun{
prior_auto()
}
}
