% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prior_manual.R
\name{prior_manual}
\alias{prior_manual}
\title{1D Prior Elicitation with manual features}
\usage{
prior_manual()
}
\value{
opens shiny app
}
\description{
1D Prior Elicitation with manual features
}
\examples{
\dontrun{
prior_manual()
}
}
