% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_nimble.R
\name{run_nimble}
\alias{run_nimble}
\title{run_nimble}
\usage{
run_nimble(
  Y,
  X_mat,
  code_in,
  constants_in,
  data_in,
  tau_in,
  sigma_in = 1,
  num_sims = 6000,
  num_burn = 1000
)
}
\arguments{
\item{Y}{response vector}

\item{X_mat}{matrix of covariates}

\item{code_in}{list with nimble code}

\item{constants_in}{list with constants for nimble code}

\item{data_in}{list with data for nimble code}

\item{tau_in}{prior variance on betas}

\item{sigma_in}{fixed value of sigma}

\item{num_sims}{number of MCMC sims}

\item{num_burn}{number of iterations for burn in}
}
\value{
list with mcmc simulations generated in nimble
}
\description{
run_nimble
}
\examples{
\dontrun{
run_nimble()
}
}
